refactor(gdrive): extract lock-management helpers into gdrive_locking module (#1137)
Move _read_lockfile_metadata, _print_lockfile_messages, _check_pid_alive,
and _acquire_or_bypass_lock from gdrive_cli.py into a new gdrive_locking.py
module. Each helper accepts its collaborators (tool, console, args) explicitly
so the new module has no circular dependency back into gdrive_cli.

- gdrive_cli.py: remove the four definitions, add import from gdrive_locking,
  and drop the now-unused `time` import.
- test_gdrive_locking.py: full coverage of all lock paths (stale-lock,
  live-lock, --force bypass, wait-loop integer/fractional display, timeout
  expiry, acquired-on-retry); patches target gdrive_locking.* not gdrive_cli.*.
- test_gdrive_cli_lock_management.py: retain _run_and_release_lock and
  _apply_retry_failed_file tests; remove tests for moved functions.
- CHANGELOG and VERSION bumped to 1.30.0.
