+++ src/python/cloud/gdrive_constants.py
@@ -12,7 +12,7 @@ import os
 # ---------------------------------------------------------------------------
 # Version
 # ---------------------------------------------------------------------------
-VERSION = "1.29.0"
+VERSION = "1.30.0"
 
 # ---------------------------------------------------------------------------
 # CLI help text
